+++ stanza/models/tokenizer.py
@@ -13,9 +13,9 @@ import numpy as np
 import torch
 
 from stanza.models.common import utils
-from stanza.models.tokenize.trainer import Trainer
-from stanza.models.tokenize.data import DataLoader
-from stanza.models.tokenize.utils import load_mwt_dict, eval_model, output_predictions
+from stanza.models.tokenization.trainer import Trainer
+from stanza.models.tokenization.data import DataLoader
+from stanza.models.tokenization.utils import load_mwt_dict, eval_model, output_predictions
 from stanza.models import _training_logging
 
 def parse_args():
